tests/wpt/tests/css/css-text/parsing/text-spacing-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-spacing with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-spacing-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// These keywords must appear alone.
for (const keyword of ['none', 'auto']) {
  test_invalid_value("text-spacing", `${keyword} ${keyword}`);
  for (const keyword2 of ['no-autospace', 'space-first']) {
    test_invalid_value("text-spacing", `${keyword} ${keyword2}`);
    test_invalid_value("text-spacing", `${keyword2} ${keyword}`);
  }
}
</script>
</body>
</html>
